Move selected state back to RequestTable
